pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["Diego França <diego.franca92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
gunicorn = "^21.0.1"
whitenoise = "^6.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.3"
django-cors-headers = "^4.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
